backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

python-multipart==0.0.9
markdown==3.7

google-genai==1.*
pillow==10.*
argon2-cffi==23.1.0
